chore: move to own demo page
