+++ src/app/app.component.html
@@ -31,6 +31,7 @@
             <a class="list-group-item" routerLink="/tabs" routerLinkActive="active">Tabs</a>
             <a class="list-group-item" routerLink="/toasts" routerLinkActive="active">Toasts</a>
             <a class="list-group-item" routerLink="/tooltips" routerLinkActive="active">Tooltips</a>
+            <a class="list-group-item" routerLink="/word-cloud" routerLinkActive="active">Word Cloud</a>
         </div>
     </div>
 </div>
