--- src/app/home/home.component.scss
@@ -1,6 +1,0 @@
-.word-cloud-container {
-    width: 500px;
-    height: 300px;
-    border: 1px solid rgb(72, 255, 0);
-    margin-top: 90px;
-}
